+++ app/src/main/java/com/github/sdpteam15/polyevents/model/map/ZoneAreaMapHelper.kt
@@ -1,0 +1,4 @@
+package com.github.sdpteam15.polyevents.model.map
+
+object ZoneAreaMapHelper {
+}
