+++ app/src/test/java/com/github/sdpteam15/polyevents/map/GoogleMapHelperRotationsTests.kt
@@ -1,6 +1,7 @@
 package com.github.sdpteam15.polyevents.map
 
 import com.github.sdpteam15.polyevents.model.map.GoogleMapHelper
+import com.github.sdpteam15.polyevents.model.map.GoogleMapVectorHelper
 import com.google.android.gms.maps.model.LatLng
 import org.hamcrest.MatcherAssert.assertThat
 import org.hamcrest.Matchers.closeTo
@@ -20,8 +21,8 @@ class GoogleMapHelperRotationsTests {
         val center = LatLng(46.52100506978624, 6.565499156713487)
         val point = LatLng(46.52111073013754, 6.565624214708805)
 
-        val projectedPoint = GoogleMapHelper.equirectangularProjection(point, center)
-        val transformedPoint = GoogleMapHelper.inverseEquirectangularProjection(
+        val projectedPoint = GoogleMapVectorHelper.equirectangularProjection(point, center)
+        val transformedPoint = GoogleMapVectorHelper.inverseEquirectangularProjection(
             projectedPoint,
             center
         )
@@ -41,14 +42,14 @@ class GoogleMapHelperRotationsTests {
             (points[0].longitude + points[1].longitude) / 2
         )
 
-        var center = GoogleMapHelper.getCenter(points)
+        var center = GoogleMapVectorHelper.getCenter(points)
 
         assertLatLng(center, middle)
 
         points.add(LatLng(30.02938498, 45.20934809))
 
         middle = LatLng(40.93120003348288, 18.736479604616118)
-        center = GoogleMapHelper.getCenter(points)
+        center = GoogleMapVectorHelper.getCenter(points)
 
         assertLatLng(center, middle)
     }
@@ -58,10 +59,10 @@ class GoogleMapHelperRotationsTests {
         val p0 = LatLng(46.52100506978624, 6.565499156713487)
         val p1 = LatLng(46.2432100506624, 4.43459156713487)
 
-        val center = GoogleMapHelper.getCenter(listOf(p0, p1))
+        val center = GoogleMapVectorHelper.getCenter(listOf(p0, p1))
 
-        val p0c = GoogleMapHelper.equirectangularProjection(p0, center)
-        val p1c = GoogleMapHelper.equirectangularProjection(p1, center)
+        val p0c = GoogleMapVectorHelper.equirectangularProjection(p0, center)
+        val p1c = GoogleMapVectorHelper.equirectangularProjection(p1, center)
 
         val distance =
             sqrt((p0c.first - p1c.first) * (p0c.first - p1c.first) + (p0c.second - p1c.second) * (p0c.second - p1c.second))
@@ -70,23 +71,23 @@ class GoogleMapHelperRotationsTests {
 
     @Test
     fun degreeToRadianIsCorrect() {
-        assertThat(GoogleMapHelper.degreeToRadian(90.0), Is(PI / 2))
-        assertThat(GoogleMapHelper.degreeToRadian(45.0), Is(PI / 4))
+        assertThat(GoogleMapVectorHelper.degreeToRadian(90.0), Is(PI / 2))
+        assertThat(GoogleMapVectorHelper.degreeToRadian(45.0), Is(PI / 4))
     }
 
     @Test
     fun radianToDegreeIsCorrect() {
-        assertThat(GoogleMapHelper.radianToDegree(PI / 2), Is(90.0))
-        assertThat(GoogleMapHelper.radianToDegree(PI / 4), Is(45.0))
+        assertThat(GoogleMapVectorHelper.radianToDegree(PI / 2), Is(90.0))
+        assertThat(GoogleMapVectorHelper.radianToDegree(PI / 4), Is(45.0))
     }
 
     @Test
     fun getDirectionIsCorrect() {
         var p = Pair(1.0, 1.0)
-        assertThat(GoogleMapHelper.getDirection(p), Is(PI / 4))
+        assertThat(GoogleMapVectorHelper.getDirection(p), Is(PI / 4))
 
         p = Pair(-1.0, 2.0)
-        assertThat(GoogleMapHelper.getDirection(p), Is(PI - 1.1071487177940904))
+        assertThat(GoogleMapVectorHelper.getDirection(p), Is(PI - 1.1071487177940904))
     }
 
     @Test
@@ -94,12 +95,12 @@ class GoogleMapHelperRotationsTests {
         var p = Pair(3.14, 1.0)
         var angle = PI / 2
 
-        var rotated = GoogleMapHelper.computeRotation(p, angle)
+        var rotated = GoogleMapVectorHelper.computeRotation(p, angle)
         var expected = Pair(-1.0, 3.14)
         assertPoints(rotated, expected)
 
         angle = PI / 8
-        rotated = GoogleMapHelper.computeRotation(p, angle)
+        rotated = GoogleMapVectorHelper.computeRotation(p, angle)
         expected = Pair(2.5182983, 2.12550551)
         assertPoints(rotated, expected)
     }
@@ -114,7 +115,7 @@ class GoogleMapHelperRotationsTests {
         )
 
         assertThat(
-            GoogleMapHelper.computeMeanRadius(points),
+            GoogleMapVectorHelper.computeMeanRadius(points),
             Is(sqrt(1.25))
         )
     }
@@ -125,7 +126,7 @@ class GoogleMapHelperRotationsTests {
         val center = LatLng(46.52111073013754, 6.565624214708805)
         val angle = PI / 3
 
-        val result = GoogleMapHelper.applyRotation(point, angle, center)
+        val result = GoogleMapVectorHelper.applyRotation(point, angle, center)
         println(result)
         val expected = LatLng(46.520983377771145, 6.565694669641021)
         assertLatLng(result, expected)
